+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Next
+* Fix endianness issue when converting a `TitleMetadataReader` object to `bytes`
+
 ## v0.4.2 - July 28, 2020
 * Don't assume a Cygwin environment is Windows
 * Change keyslot for New 3DS key sector keys from 0x11 to 0x43